Don't create a DB with template1 using template1's owner
template1 will likely be owned by the postgres superuser, which should
be a different user then the one the app connects to the DB as.  It
won't have permission to use GRANT to give this user permissions from
the postgres superuser.

Note that they'll still fail if the user specificly requests 'owner'
to be the postgres superuser