When expiring databases, remove the record even if no database exists on the server
Tests were running slowly and the log indicated there were a bunch of
live_database records it was trying to expire, but there were no corresponding
databases on the Postgres server.  Maybe leftovers from testing.

This change catches the particular exception and goes ahead and removes
the record of this missing database.