add 'connect_db_name' attribute to PostgresInstance
It's the database name to connect to to issue commands to the database server.
PostgreSQL requires you to connect to _some_ database, even if you're going to
create other databases or add users.  We used to always connect to 'template1',
but it was messing up our server monitoring.

This adds a required attribute to specify which DB to connect to.