Return 503 if the database server can't make the database
This is intended to resolve #60, but it doesn't solve that particular problem.
It looks like the full-disk problem makes the CREATE DATABASE statement in
PostgresInstance::createdb_from_template() fail, which throws
Exception::CannotCreateDatabase.  Without parsing the test of the error
returned by the server, we can't deduce why it failed.

The test is kind if cheesy, relying on the one and only do() to be that
CREATE DATABASE statement.

We're now catching that exception and returning 503 instead of the default 400.