Change how 404s are rendered
At some point, Mojolicious deprecated and later removed the render_not_found()
(and render_exception()) methods,  The new way is to use the "reply" helper
found in Mojolicious::Plugin::DefaultHelpers like this:
    $self->reply->not_found()

That's not available in the older version we have installed currently.

This way will work with both the old and newer versions of Mojo.